# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CRM ЛГ Интеграция API
  version: 0.1.0
  description: |
    Контракт REST API CRM-системы «ЛГ Интеграция».

    Источник истины формируется в две стороны: код фронта (`frontend/src/api/*.ts`
    и `frontend/src/mocks/handlers.ts`) → OpenAPI → код бэкенда (FastAPI).

    Все мутирующие эндпоинты возвращают актуальное состояние ресурса (либо
    `{ ok: true }` для удалений). Чувствительные поля кандидатов (`email`,
    `phone`, `rateMonth`) фильтруются на сервере в зависимости от роли —
    клиент получит `null` или `"—"` вместо значения.

    WebSocket-ручки в этом контракте НЕ описаны (OpenAPI описывает HTTP-запросы,
    а не поток фреймов после Upgrade) и, соответственно, не попадают в codegen.
    Протокол комнаты AI-скрининга `/api/v1/ws/screening/{sessionId}?token=` —
    control-сообщения, формат бинарных аудио-фреймов, события сервер→клиент,
    ping/pong, коды закрытия и семантика reconnect — описан отдельно:
    `docs/ws-screening.md`.
  contact:
    name: ЛГ Интеграция
    email: ops@lg.ru
servers:
  - url: /api/v1
    description: Боевой API (через nginx reverse proxy)
  - url: http://localhost:8000/api/v1
    description: Локальная разработка
security:
  - bearerAuth: []
...
paths:
...
  /vacancies/{id}/activity:
    parameters:
      - $ref: '#/components/parameters/UuidPathId'
    get:
      tags: [vacancies, audit]
      summary: Активность по вакансии
      responses:
        '200':
          description: Activity
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ActivityEntry' }
...
  /candidates/{id}/activity:
    parameters:
      - $ref: '#/components/parameters/UuidPathId'
    get:
      tags: [candidates, audit]
      summary: Активность по кандидату
      responses:
        '200':
          description: Activity
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ActivityEntry' }
  /clients/{id}/activity:
    parameters:
      - $ref: '#/components/parameters/UuidPathId'
    get:
      tags: [clients, audit]
      summary: Активность по клиенту
      responses:
        '200':
          description: Activity
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ActivityEntry' }

  # ─────────────── COMMENTS ───────────────
...
  /audit:
    get:
      tags: [audit]
      summary: Журнал аудита с фильтрами
      parameters:
        - { in: query, name: entityType, schema: { type: string } }
        - { in: query, name: entityId,   schema: { $ref: '#/components/schemas/UUID' } }
        - { in: query, name: actorId,    schema: { $ref: '#/components/schemas/UUID' } }
        - { in: query, name: field,      schema: { type: string } }
        - { in: query, name: dateFrom,   schema: { type: string, format: date } }
        - { in: query, name: dateTo,     schema: { type: string, format: date } }
        - { in: query, name: search,     schema: { type: string } }
      responses:
        '200':
          description: Аудит
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEntry' }

  # ─────────────── ANALYTICS ───────────────
...
  /files/presign:
    post:
      tags: [files]
      summary: Получить presigned URL для загрузки файла в S3
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PresignRequest' }
      responses:
        '200':
          description: PresignResponse
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PresignResponse' }
        '422':
          description: Неподдерживаемый MIME или превышен размер (unsupported_mime / file_too_large)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /files/confirm:
    post:
      tags: [files]
      summary: Подтвердить загрузку (создаёт запись в БД, ставит scan_status=pending)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmRequest' }
      responses:
        '200':
          description: FileResponse
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileResponse' }
  /files:
    get:
      tags: [files]
      summary: Файлы сущности
      parameters:
        - { in: query, name: entityType, required: true, schema: { $ref: '#/components/schemas/FileEntityType' } }
        - { in: query, name: entityId,   required: true, schema: { $ref: '#/components/schemas/UUID' } }
      responses:
        '200':
          description: Список файлов
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FileResponse' }
  /files/{id}/download:
    parameters:
      - $ref: '#/components/parameters/UuidPathId'
    get:
      tags: [files]
      summary: Временный URL для скачивания
      responses:
        '200':
          description: DownloadResponse
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DownloadResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Файл найден антивирусом как заражённый (file_infected)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /files/{id}:
    parameters:
      - $ref: '#/components/parameters/UuidPathId'
    delete:
      tags: [files]
      summary: Удалить файл (владелец или admin)
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─────────────── DOCUMENTS ───────────────
...
  /hr-edo/documents/{id}/file:
    parameters:
      - $ref: '#/components/parameters/UuidPathId'
    get:
      tags: [hr-edo]
      summary: PDF документа (исходник с проверкой хеша или с листом подписания)
      parameters:
        - { in: query, name: kind, schema: { type: string, enum: [source, stamped], default: source } }
        - { in: query, name: download, schema: { type: boolean, default: false } }
      responses:
        '200':
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /hr-edo/documents/{id}/signatures/{signatureId}/file:
    parameters:
      - $ref: '#/components/parameters/UuidPathId'
      - { in: path, name: signatureId, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      tags: [hr-edo]
      summary: Файл подписи (.sig) или скан — hr_edo:export
      responses:
        '200':
          description: Файл
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
...
  /hr-edo/my/documents/{id}/file:
    parameters:
      - $ref: '#/components/parameters/UuidPathId'
    get:
      tags: [hr-edo]
      summary: PDF моего документа
      responses:
        '200':
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidPathId:
      in: path
      name: id
      required: true
      schema: { $ref: '#/components/schemas/UUID' }
...
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            type: object
            required: [ok]
            properties:
              ok: { type: boolean, enum: [true] }
...
    Forbidden:
      description: Нет прав
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    UUID:
      type: string
      format: uuid
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, description: 'Машиночитаемый код, например engagement_type_mismatch' }
        message: { type: string }
        details: { type: object, additionalProperties: true }

    # ─── Auth ───
...
    ActivityEntry:
      type: object
      required: [id, entityType, entityId, kind, text, createdAt]
      properties:
        id:         { $ref: '#/components/schemas/UUID' }
        entityType:
          type: string
          enum: [vacancy, candidate, client]
        entityId:   { $ref: '#/components/schemas/UUID' }
        # Может быть null: актёр удалён.
        actorId:    { allOf: [{ $ref: '#/components/schemas/UUID' }], nullable: true }
        kind:
          type: string
          enum: [create, status, note, call, email]
        text:       { type: string }
        createdAt:  { type: string, format: date-time }
    AuditEntry:
      type: object
      required: [id, entityType, entityId, field, before, after, createdAt]
      properties:
        id:         { $ref: '#/components/schemas/UUID' }
        entityType: { type: string }
        entityId:   { $ref: '#/components/schemas/UUID' }
        # Может быть null: актёр удалён.
        actorId:    { allOf: [{ $ref: '#/components/schemas/UUID' }], nullable: true }
        field:      { type: string }
        before:     { type: string, nullable: true }
        after:      { type: string, nullable: true }
        createdAt:  { type: string, format: date-time }

    # ─── Analytics ───
...
    FileEntityType:
      type: string
      enum: [candidate, vacancy, client, contact, document, hr_document]
    ScanStatus:
      type: string
      enum: [pending, clean, infected, error]
    FileResponse:
      type: object
      required: [id, entityType, entityId, fileKey, originalName, mime, size, scanStatus, createdAt]
      properties:
        id:           { $ref: '#/components/schemas/UUID' }
        # Может быть null: владелец удалён.
        ownerUserId:  { allOf: [{ $ref: '#/components/schemas/UUID' }], nullable: true }
        entityType:   { $ref: '#/components/schemas/FileEntityType' }
        entityId:     { $ref: '#/components/schemas/UUID' }
        fileKey:      { type: string }
        originalName: { type: string }
        mime:         { type: string }
        size:         { type: integer, format: int64 }
        scanStatus:   { $ref: '#/components/schemas/ScanStatus' }
        scannedAt:    { type: string, format: date-time, nullable: true }
        createdAt:    { type: string, format: date-time }
    PresignRequest:
      type: object
      required: [entityType, entityId, originalName, mime, size]
      properties:
        entityType:   { $ref: '#/components/schemas/FileEntityType' }
        entityId:     { $ref: '#/components/schemas/UUID' }
        originalName: { type: string, minLength: 1, maxLength: 512 }
        mime:         { type: string }
        size:         { type: integer, minimum: 1 }
    PresignResponse:
      type: object
      required: [url, fields, fileKey, maxBytes]
      properties:
        url:      { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
        fileKey:  { type: string }
        maxBytes: { type: integer }
    ConfirmRequest:
      type: object
      required: [fileKey, entityType, entityId, originalName, mime, size]
      properties:
        fileKey:      { type: string }
        entityType:   { $ref: '#/components/schemas/FileEntityType' }
        entityId:     { $ref: '#/components/schemas/UUID' }
        originalName: { type: string }
        mime:         { type: string }
        size:         { type: integer, minimum: 1 }
    DownloadResponse:
      type: object
      required: [url, expiresIn]
      properties:
        url:       { type: string }
        expiresIn: { type: integer }

    # ─── Documents ───
...